pyproject.toml
[project]
name = "gd-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.3",
]
